Add rootWindow/errors module
